Refactor PAL into a namespace with separate header files
