+++ include/ccf/endpoint_registry.h
@@ -3,9 +3,9 @@
 #pragma once
 
 #include "ccf/ds/json_schema.h"
-#include "ccf/ds/pal.h"
 #include "ccf/endpoint.h"
 #include "ccf/endpoint_context.h"
+#include "ccf/pal/locking.h"
 #include "ccf/rpc_context.h"
 #include "ccf/tx.h"
 
@@ -127,7 +127,7 @@ namespace ccf::endpoints
       std::map<RESTVerb, std::shared_ptr<PathTemplatedEndpoint>>>
       templated_endpoints;
 
-    ccf::Pal::Mutex metrics_lock;
+    ccf::pal::Mutex metrics_lock;
     std::map<std::string, std::map<std::string, Metrics>> metrics;
 
     EndpointRegistry::Metrics& get_metrics_for_endpoint(
