+++ src/consensus/aft/raft.h
@@ -4,7 +4,7 @@
 
 #include "ccf/ds/ccf_exception.h"
 #include "ccf/ds/logger.h"
-#include "ccf/ds/pal.h"
+#include "ccf/pal/locking.h"
 #include "ccf/tx_id.h"
 #include "ccf/tx_status.h"
 #include "ds/serialized.h"
@@ -245,7 +245,7 @@ namespace aft
 
     bool view_change_in_progress() override
     {
-      std::unique_lock<ccf::Pal::Mutex> guard(state->lock);
+      std::unique_lock<ccf::pal::Mutex> guard(state->lock);
       if (consensus_type == ConsensusType::BFT)
       {
         LOG_FAIL_FMT("Unsupported");
@@ -274,13 +274,13 @@ namespace aft
 
     bool can_replicate() override
     {
-      std::unique_lock<ccf::Pal::Mutex> guard(state->lock);
+      std::unique_lock<ccf::pal::Mutex> guard(state->lock);
       return can_replicate_unsafe();
     }
 
     Consensus::SignatureDisposition get_signature_disposition() override
     {
-      std::unique_lock<ccf::Pal::Mutex> guard(state->lock);
+      std::unique_lock<ccf::pal::Mutex> guard(state->lock);
       if (can_replicate_unsafe())
       {
         if (should_sign)
@@ -328,7 +328,7 @@ namespace aft
     {
       // When receiving append entries as a follower, all security domains will
       // be deserialised
-      std::lock_guard<ccf::Pal::Mutex> guard(state->lock);
+      std::lock_guard<ccf::pal::Mutex> guard(state->lock);
       public_only = false;
     }
 
@@ -347,7 +347,7 @@ namespace aft
           "Can't force leadership if there is already a leader");
       }
 
-      std::lock_guard<ccf::Pal::Mutex> guard(state->lock);
+      std::lock_guard<ccf::pal::Mutex> guard(state->lock);
       state->current_view += starting_view_change;
       become_leader(true);
     }
@@ -366,7 +366,7 @@ namespace aft
           "Can't force leadership if there is already a leader");
       }
 
-      std::lock_guard<ccf::Pal::Mutex> guard(state->lock);
+      std::lock_guard<ccf::pal::Mutex> guard(state->lock);
       state->current_view = term;
       state->last_idx = index;
       state->commit_idx = commit_idx_;
@@ -384,7 +384,7 @@ namespace aft
     {
       // This should only be called when the node resumes from a snapshot and
       // before it has received any append entries.
-      std::lock_guard<ccf::Pal::Mutex> guard(state->lock);
+      std::lock_guard<ccf::pal::Mutex> guard(state->lock);
 
       state->last_idx = index;
       state->commit_idx = index;
@@ -403,26 +403,26 @@ namespace aft
 
     Index get_committed_seqno() override
     {
-      std::lock_guard<ccf::Pal::Mutex> guard(state->lock);
+      std::lock_guard<ccf::pal::Mutex> guard(state->lock);
       return get_commit_idx_unsafe();
     }
 
     Term get_view() override
     {
-      std::lock_guard<ccf::Pal::Mutex> guard(state->lock);
+      std::lock_guard<ccf::pal::Mutex> guard(state->lock);
       return state->current_view;
     }
 
     std::pair<Term, Index> get_committed_txid() override
     {
-      std::lock_guard<ccf::Pal::Mutex> guard(state->lock);
+      std::lock_guard<ccf::pal::Mutex> guard(state->lock);
       ccf::SeqNo commit_idx = get_commit_idx_unsafe();
       return {get_term_internal(commit_idx), commit_idx};
     }
 
     kv::Consensus::SignableTxIndices get_signable_txid() override
     {
-      std::lock_guard<ccf::Pal::Mutex> guard(state->lock);
+      std::lock_guard<ccf::pal::Mutex> guard(state->lock);
 
       kv::Consensus::SignableTxIndices r;
       r.version = state->last_idx;
@@ -438,7 +438,7 @@ namespace aft
 
     Term get_view(Index idx) override
     {
-      std::lock_guard<ccf::Pal::Mutex> guard(state->lock);
+      std::lock_guard<ccf::pal::Mutex> guard(state->lock);
       return get_term_internal(idx);
     }
 
@@ -635,7 +635,7 @@ namespace aft
     std::optional<kv::Configuration::Nodes> orc(
       kv::ReconfigurationId rid, const ccf::NodeId& node_id) override
     {
-      std::lock_guard<ccf::Pal::Mutex> guard(state->lock);
+      std::lock_guard<ccf::pal::Mutex> guard(state->lock);
 
       LOG_DEBUG_FMT(
         "Configurations: ORC for configuration #{} from {}", rid, node_id);
@@ -699,14 +699,14 @@ namespace aft
 
     Configuration::Nodes get_latest_configuration() override
     {
-      std::lock_guard<ccf::Pal::Mutex> guard(state->lock);
+      std::lock_guard<ccf::pal::Mutex> guard(state->lock);
       return get_latest_configuration_unsafe();
     }
 
     kv::ConsensusDetails get_details() override
     {
       kv::ConsensusDetails details;
-      std::lock_guard<ccf::Pal::Mutex> guard(state->lock);
+      std::lock_guard<ccf::pal::Mutex> guard(state->lock);
       details.primary_id = leader_id;
       details.current_view = state->current_view;
       details.ticking = ticking;
@@ -735,7 +735,7 @@ namespace aft
 
     bool replicate(const kv::BatchVector& entries, Term term) override
     {
-      std::lock_guard<ccf::Pal::Mutex> guard(state->lock);
+      std::lock_guard<ccf::pal::Mutex> guard(state->lock);
 
       if (leadership_state != kv::LeadershipState::Leader)
       {
@@ -895,7 +895,7 @@ namespace aft
 
     void periodic(std::chrono::milliseconds elapsed) override
     {
-      std::unique_lock<ccf::Pal::Mutex> guard(state->lock);
+      std::unique_lock<ccf::pal::Mutex> guard(state->lock);
       timeout_elapsed += elapsed;
 
       if (leadership_state == kv::LeadershipState::Leader)
@@ -1146,7 +1146,7 @@ namespace aft
       const uint8_t* data,
       size_t size)
     {
-      std::unique_lock<ccf::Pal::Mutex> guard(state->lock);
+      std::unique_lock<ccf::pal::Mutex> guard(state->lock);
 
       LOG_DEBUG_FMT(
         "Received append entries: {}.{} to {}.{} (from {} in term {})",
@@ -1521,7 +1521,7 @@ namespace aft
     void recv_append_entries_response(
       const ccf::NodeId& from, AppendEntriesResponse r)
     {
-      std::lock_guard<ccf::Pal::Mutex> guard(state->lock);
+      std::lock_guard<ccf::pal::Mutex> guard(state->lock);
       // Ignore if we're not the leader.
 
       if (leadership_state != kv::LeadershipState::Leader)
@@ -1655,7 +1655,7 @@ namespace aft
 
     void recv_request_vote(const ccf::NodeId& from, RequestVote r)
     {
-      std::lock_guard<ccf::Pal::Mutex> guard(state->lock);
+      std::lock_guard<ccf::pal::Mutex> guard(state->lock);
 
       // Do not check that from is a known node. It is possible to receive
       // RequestVotes from nodes that this node doesn't yet know, just as it
@@ -1763,7 +1763,7 @@ namespace aft
     void recv_request_vote_response(
       const ccf::NodeId& from, RequestVoteResponse r)
     {
-      std::lock_guard<ccf::Pal::Mutex> guard(state->lock);
+      std::lock_guard<ccf::pal::Mutex> guard(state->lock);
 
       if (leadership_state != kv::LeadershipState::Candidate)
       {
