+++ src/enclave/enclave.h
@@ -3,7 +3,8 @@
 #pragma once
 #include "ccf/app_interface.h"
 #include "ccf/ds/logger.h"
-#include "ccf/ds/pal.h"
+#include "ccf/pal/enclave.h"
+#include "ccf/pal/mem.h"
 #include "ds/oversized.h"
 #include "enclave_time.h"
 #include "indexing/enclave_lfs_access.h"
@@ -90,7 +91,7 @@ namespace ccf
       rpc_map(std::make_shared<RPCMap>()),
       rpcsessions(std::make_shared<RPCSessions>(*writer_factory, rpc_map))
     {
-      ccf::Pal::initialize_enclave();
+      ccf::pal::initialize_enclave();
       ccf::initialize_verifiers();
 
       // From
@@ -189,7 +190,7 @@ namespace ccf
       }
       LOG_TRACE_FMT("Shutting down enclave");
       ccf::shutdown_verifiers();
-      ccf::Pal::shutdown_enclave();
+      ccf::pal::shutdown_enclave();
     }
 
     CreateNodeStatus create_new_node(
@@ -232,7 +233,7 @@ namespace ccf
           r.self_signed_node_cert.size());
         return CreateNodeStatus::InternalError;
       }
-      Pal::safe_memcpy(
+      pal::safe_memcpy(
         node_cert,
         r.self_signed_node_cert.data(),
         r.self_signed_node_cert.size());
@@ -250,7 +251,7 @@ namespace ccf
             r.service_cert.size());
           return CreateNodeStatus::InternalError;
         }
-        Pal::safe_memcpy(
+        pal::safe_memcpy(
           service_cert, r.service_cert.data(), r.service_cert.size());
         *service_cert_len = r.service_cert.size();
       }
