+++ src/endpoints/endpoint_registry.cpp
@@ -435,28 +435,28 @@ namespace ccf::endpoints
 
   void EndpointRegistry::increment_metrics_calls(const EndpointDefinitionPtr& e)
   {
-    std::lock_guard<ccf::Pal::Mutex> guard(metrics_lock);
+    std::lock_guard<ccf::pal::Mutex> guard(metrics_lock);
     get_metrics_for_endpoint(e).calls++;
   }
 
   void EndpointRegistry::increment_metrics_errors(
     const EndpointDefinitionPtr& e)
   {
-    std::lock_guard<ccf::Pal::Mutex> guard(metrics_lock);
+    std::lock_guard<ccf::pal::Mutex> guard(metrics_lock);
     get_metrics_for_endpoint(e).errors++;
   }
 
   void EndpointRegistry::increment_metrics_failures(
     const EndpointDefinitionPtr& e)
   {
-    std::lock_guard<ccf::Pal::Mutex> guard(metrics_lock);
+    std::lock_guard<ccf::pal::Mutex> guard(metrics_lock);
     get_metrics_for_endpoint(e).failures++;
   }
 
   void EndpointRegistry::increment_metrics_retries(
     const EndpointDefinitionPtr& e)
   {
-    std::lock_guard<ccf::Pal::Mutex> guard(metrics_lock);
+    std::lock_guard<ccf::pal::Mutex> guard(metrics_lock);
     get_metrics_for_endpoint(e).retries++;
   }
 }
