+++ src/host/tcp.h
@@ -101,7 +101,7 @@ namespace asynchost
     ~TCPImpl()
     {
       {
-        std::unique_lock<ccf::Pal::Mutex> guard(pending_resolve_requests_mtx);
+        std::unique_lock<ccf::pal::Mutex> guard(pending_resolve_requests_mtx);
         for (auto& req : pending_resolve_requests)
         {
           // The UV request objects can stay, but if there are any references
@@ -562,7 +562,7 @@ namespace asynchost
 
     static void on_resolved(uv_getaddrinfo_t* req, int rc, struct addrinfo* res)
     {
-      std::unique_lock<ccf::Pal::Mutex> guard(pending_resolve_requests_mtx);
+      std::unique_lock<ccf::pal::Mutex> guard(pending_resolve_requests_mtx);
       pending_resolve_requests.erase(req);
 
       if (req->data)
