+++ src/host/udp.h
@@ -102,7 +102,7 @@ namespace asynchost
     ~UDPImpl()
     {
       {
-        std::unique_lock<ccf::Pal::Mutex> guard(pending_resolve_requests_mtx);
+        std::unique_lock<ccf::pal::Mutex> guard(pending_resolve_requests_mtx);
         for (auto& req : pending_resolve_requests)
         {
           // The UV request objects can stay, but if there are any references
@@ -353,7 +353,7 @@ namespace asynchost
 
     static void on_resolved(uv_getaddrinfo_t* req, int rc, struct addrinfo* res)
     {
-      std::unique_lock<ccf::Pal::Mutex> guard(pending_resolve_requests_mtx);
+      std::unique_lock<ccf::pal::Mutex> guard(pending_resolve_requests_mtx);
       pending_resolve_requests.erase(req);
 
       LOG_TRACE_FMT("UDP on_resolve static");
