+++ src/kv/untyped_map.h
@@ -3,8 +3,8 @@
 #pragma once
 
 #include "ccf/ds/logger.h"
-#include "ccf/ds/pal.h"
 #include "ccf/kv/untyped_map_handle.h"
+#include "ccf/pal/locking.h"
 #include "ds/dl_list.h"
 #include "kv/kv_serialiser.h"
 #include "kv/kv_types.h"
@@ -82,7 +82,7 @@ namespace kv::untyped
     CommitHook global_hook = nullptr;
     MapHook hook = nullptr;
     std::list<std::pair<Version, Write>> commit_deltas;
-    ccf::Pal::Mutex sl;
+    ccf::pal::Mutex sl;
     const SecurityDomain security_domain;
     const bool replicated;
     const bool include_conflict_read_version;
